pyproject.toml
[tool.poetry]
name = "qr-codes"
version = "0.1.0"
description = "Utility to generate QR codes"
authors = ["makerbynight <2565361+mshafer1@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
qr-code = "qr_codes.__main__:_cli"

[tool.poetry.dependencies]
python = "^3.9"
qrcode = {extras = ["pillow"], version = "^8.2"}
pillow = "^11.3.0"
click = "^8.1.7"
click-option-group = "^0.5.7"


[tool.poetry.group.dev.dependencies]
ni-python-styleguide = "^0.4.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
